decant listeners to own mod and remove ext
