+++ pombot/__init__.py
@@ -1,1 +1,2 @@
 from pombot import commands
+from pombot import listeners
