+++ pombot/extensions/general.py
@@ -1,3 +1,5 @@
+from functools import partial
+
 from discord.ext.commands import Command
 from discord.ext.commands.bot import Bot
 
@@ -6,6 +8,14 @@ import pombot
 
 def setup(bot: Bot):
     """Load general, uncategorized commands and listeners."""
+    listeners = [
+        (partial(pombot.listeners.handle_on_ready, bot), "on_ready"),
+        (pombot.listeners.handle_on_command_error,       "on_command_error"),
+    ]
+
+    for listener in listeners:
+        bot.add_listener(*listener)
+
     commands = [
         Command(pombot.commands.do_events,  name="events"),
         Command(pombot.commands.do_howmany, name="howmany"),
