+++ pombot/listeners/__init__.py
@@ -1,0 +1,2 @@
+from pombot.listeners.on_command_error import *
+from pombot.listeners.on_ready import *
